fixup: single test suite
