+++ providers/flagd/src/test/java/dev/openfeature/contrib/providers/flagd/e2e/process/FlagdInProcessSetup.java
@@ -1,0 +1,24 @@
+package dev.openfeature.contrib.providers.flagd.e2e.process;
+
+import dev.openfeature.contrib.providers.flagd.Config;
+import dev.openfeature.contrib.providers.flagd.FlagdOptions;
+import dev.openfeature.contrib.providers.flagd.FlagdProvider;
+import dev.openfeature.contrib.providers.flagd.e2e.StepDefinitions;
+import dev.openfeature.sdk.Client;
+import dev.openfeature.sdk.OpenFeatureAPI;
+import io.cucumber.java.BeforeAll;
+
+public class FlagdInProcessSetup {
+    
+    @BeforeAll()
+    public static void setup() throws InterruptedException {
+        FlagdProvider provider = new FlagdProvider(FlagdOptions.builder()
+        .resolverType(Config.Evaluator.IN_PROCESS)
+        .host("localhost")
+        .port(9090)
+        .build());
+        OpenFeatureAPI.getInstance().setProviderAndWait("process", provider);
+        Client client = OpenFeatureAPI.getInstance().getClient("process");
+        StepDefinitions.setClient(client);
+    }
+}
